WIP: partial package extraction (agent died on 529 overload)
Durability commit -- the agent was terminated mid-work by a transient
server-side 529, not a logic error. Committing its work-in-progress so it
survives worktree teardown and can be resumed. NOT complete, NOT for merge:
git mv done, shims/import-fixes likely partial. Verify before building on it.
