+++ examples/canonicalizer_enrichment.py
@@ -27,8 +27,8 @@ from pathlib import Path
 from typing import Any
 
 from langres.core import Resolver
-from langres.core.anchor_store import AnchorStore
-from langres.core.canonicalizer import Canonicalizer
+from langres.curation.anchor_store import AnchorStore
+from langres.curation.canonicalizer import Canonicalizer
 from langres.core.models import CompanySchema
 
 # ----------------------------------------------------------------------------
