+++ examples/flywheel_closed_loop.py
@@ -66,7 +66,7 @@ from pydantic import BaseModel
 from langres.clients.openrouter import BudgetExceeded, SpendMonitor
 from langres.core.calibration import derive_threshold
 from langres.core.comparators import StringComparator
-from langres.core.harvest import (
+from langres.curation.harvest import (
     Correction,
     CorrectionLog,
     LabeledPair,
@@ -79,7 +79,7 @@ from langres.core.matcher import Matcher
 from langres.core.matchers.cascade_judge import CASCADE_ESCALATED_STEP, CascadeMatcher
 from langres.core.matchers.random_forest_judge import RandomForestMatcher
 from langres.core.reports import ScoreInspectionReport, _inspect_scores_impl
-from langres.core.review import ReviewQueue, select_for_review
+from langres.curation.review import ReviewQueue, select_for_review
 
 #: Committed fixtures (regenerate with ``data/flywheel_loop/generate_fixtures.py``).
 DATA_DIR = Path(__file__).resolve().parent / "data" / "flywheel_loop"
