+++ examples/incremental_assign.py
@@ -18,7 +18,7 @@ import tempfile
 from pathlib import Path
 
 from langres.core import CompanySchema, Resolver
-from langres.core.anchor_store import AnchorStore
+from langres.curation.anchor_store import AnchorStore
 
 # A prior batch: records 1 and 2 are the same company; 3 and 4 are unique.
 BATCH = [
