+++ src/langres/core/_exports/_flywheel.py
@@ -3,7 +3,7 @@
 See ``langres.core._exports`` for the fragment contract.
 """
 
-from langres.core.harvest import (
+from langres.curation.harvest import (
     Correction,
     CorrectionLog,
     LabeledPair,
@@ -11,7 +11,7 @@ from langres.core.harvest import (
     harvest_labeled_pairs,
 )
 from langres.core.judgement_log import JudgementLog, LoggingMatcher
-from langres.core.review import ReviewItem, ReviewQueue, select_for_review
+from langres.curation.review import ReviewItem, ReviewQueue, select_for_review
 
 __all__ = [
     "Correction",
