+++ src/langres/core/_exports/_training.py
@@ -11,7 +11,7 @@ from typing import TYPE_CHECKING
 
 from langres.core.fit import SupervisedFitMixin, UnsupervisedFitMixin
 from langres.core.fit_report import FitReport
-from langres.core.harvest import align_pairs
+from langres.curation.harvest import align_pairs
 from langres.core.methods_api import Method, UnsupportedMethodKind
 from langres.core.methods_calibrate import Isotonic, Platt
 from langres.core.methods_prompt import Bootstrap, GEPA, MIPRO
