+++ src/langres/core/_model_state.py
@@ -100,7 +100,7 @@ class ModelState:
 
                 Scope, precisely: the cap meters **every seam that scores
                 through the matcher** -- ``resolve``, ``predict``, ``fit``, and
-                :meth:`~langres.core.anchor_store.AnchorStore.assign` -- across
+                :meth:`~langres.curation.anchor_store.AnchorStore.assign` -- across
                 every call on this instance, because they all route through
                 ``_scorer``. It bounds spend at ``budget_usd`` plus at most
                 one further call (see :mod:`langres.core.spend_cap`).
