+++ src/langres/curation/_pairs.py
