+++ tests/core/test_anchor_store.py
@@ -29,10 +29,10 @@ from langres.core import (
     PairwiseJudgement,
     Resolver,
 )
-from langres.core.anchor_store import AnchorStore, ClusterDelta
+from langres.curation.anchor_store import AnchorStore, ClusterDelta
 from langres.core.blockers import AllPairsBlocker, CompositeBlocker, KeyBlocker
 from langres.core.matchers import EmbeddingScoreMatcher, WeightedAverageMatcher
-from langres.core.anchor_store import _schema_factory
+from langres.curation.anchor_store import _schema_factory
 from langres.core.blockers.vector import VectorBlocker
 from langres.core.embeddings import FakeEmbedder
 from langres.core.indexes.vector_index import FAISSIndex
