+++ tests/test_flywheel_integration.py
@@ -20,7 +20,7 @@ import pytest
 
 from langres.core.judgement_log import JudgementLog
 from langres.core.models import PairwiseJudgement
-from langres.core.review import select_for_review
+from langres.curation.review import select_for_review
 
 
 def _decider(
