+++ tests/test_import_budget.py
@@ -629,8 +629,8 @@ _MOVED_OFF_THE_FACADE: dict[str, str] = {
     "KeyBlocker": "langres.core.blockers",
     "VectorBlocker": "langres.core.blockers",
     "StringComparator": "langres.core.comparators",
-    "AnchorStore": "langres.core.anchor_store",
-    "Canonicalizer": "langres.core.canonicalizer",
+    "AnchorStore": "langres.curation.anchor_store",
+    "Canonicalizer": "langres.curation.canonicalizer",
     "CorrelationClusterer": "langres.core.clusterers",
     "CascadeMatcher": "langres.core.matchers",
     "EmbeddingScoreMatcher": "langres.core.matchers",
